% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo-watermark.R
\name{ojo_watermark}
\alias{ojo_watermark}
\title{Watermark}
\usage{
ojo_watermark(
  text = "DRAFT",
  fontsize = 120,
  color = "grey90",
  alpha = 0.1,
  fontface = "bold",
  angle = 22
)
}
\arguments{
\item{text}{String to be added as watermark}

\item{fontsize}{Font size}

\item{color}{Font color}

\item{alpha}{Alpha (transparency; lower number = more transparent)}

\item{fontface}{Font face ("bold" by default)}

\item{angle}{Angle of the watermark}
}
\description{
Add a watermark to a ggplot2 plot
}
\examples{

library(ojodb)
library(ggplot2)
library(grid)

ojo_tbl("case") |>
 head(10) |>
 ojo_collect() |>
 ggplot(aes(x = date_filed, y = date_closed)) +
   geom_point() +
   ojo_theme() +
   ojo_watermark("DRAFT")

}
